Remove control_toolbox dependency and fix sim_radio warning

+++ radio/ateam_ssl_simulation_radio_bridge/src/ssl_simulation_radio_bridge_node.cpp
@@ -175,7 +175,7 @@ public:
       return;
     }
 
-    if (robot_id >= world_.our_robots.size()) {
+    if (robot_id >= static_cast<int>(world_.our_robots.size())) {
       return;
     }
